requirements.txt
python-magic-bin==0.4.14
Pillow
PyPDF2
python-docx
mutagen